requirements.txt
aiohttp
pyTelegramBotAPI
python-dotenv
pyChatGPT
PyVirtualDisplay
